Very rough non-working version of this
This has a few simple transducers for working with datasnapshots, which
works fine.

Then it has a few attempts at even trickier stuff, which does not.

Ideally I'd like to be able to extract data from multiple transducers
to compose a React element, but I can't really figure out how to do
multiple transducer passes in a single transducer. My mind is not
strong enough.